refactor(coherence): extract cvg coherence suite into convergio-coherence crate
`convergio-cli` was at 11000 / 11000 lines (CONSTITUTION § 13 hard
cap) so the deferred ADR-status verifier work (PR #144, ~525 lines)
could not land. Per CONSTITUTION § 13 we split rather than bump the
cap.

Move the seven `commands/coherence*.rs` files (1422 lines) into a
new cross-cutting crate `convergio-coherence`. The CLI keeps a thin
shim that re-exports `CoherenceCommand` and dispatches to the new
crate's `run`, converting the CLI's `OutputMode` into the crate's
mirror enum.

Behavioural no-op:
- `cvg coherence routes --output json` is byte-identical pre/post.
- `cvg coherence check --output json` violations list is identical;
  only the counters differ (one new ADR, one new crate, three new
  AGENTS.md / CLAUDE.md / README.md docs from the new crate).
- All 22 coherence tests move with their modules and pass without
  modification.

Post-state line counts:
- convergio-cli: 11000 -> 9600 lines
- convergio-coherence: new, 1466 lines

Adds ADR-0040 capturing the rationale (mirrors ADR-0013 for
durability). AUTO blocks in AGENTS.md, copilot-instructions, and
docs/adr/README.md regenerated via `cvg docs regenerate`.

Convergio plan: ee9ab55a-eb28-41ee-bd02-e22d002e8b3c
Convergio task: 76fa5d30-2e98-471a-b468-18a72f942db4
