+++ crates/convergio-cli/AGENTS.md
@@ -19,14 +19,13 @@ The block below is rewritten by `cvg docs regenerate` (ADR-0015) —
 do not edit between the markers.
 
 <!-- BEGIN AUTO:crate_stats -->
-**`convergio-cli` stats:** 63 `*.rs` files / 49 public items / 10306 lines (under `src/`).
+**`convergio-cli` stats:** 57 `*.rs` files / 48 public items / 8906 lines (under `src/`).
 
 Files approaching the 300-line cap:
 - `src/main.rs` (296 lines)
 - `src/commands/update_repo_root.rs` (292 lines)
 - `src/commands/graph.rs` (288 lines)
 - `src/commands/service.rs` (288 lines)
-- `src/commands/coherence_routes_parse.rs` (278 lines)
 - `src/commands/status_render.rs` (272 lines)
 - `src/commands/update_run.rs` (272 lines)
 - `src/commands/setup.rs` (271 lines)
