+++ crates/convergio-cli/src/commands/mod.rs
@@ -9,12 +9,6 @@ pub mod bus;
 pub mod capability;
 mod capability_types;
 pub mod coherence;
-mod coherence_body;
-mod coherence_body_scan;
-mod coherence_parse;
-mod coherence_routes;
-mod coherence_routes_diff;
-mod coherence_routes_parse;
 pub mod crdt;
 pub mod dash;
 pub mod demo;
