+++ crates/convergio-coherence/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "convergio-coherence"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Documentation/code coherence verifiers for Convergio (`cvg coherence` suite)"
+readme = "README.md"
+keywords = ["coherence", "convergio", "verifier"]
+categories = ["development-tools"]
+
+[lints]
+workspace = true
+
+[dependencies]
+convergio-i18n = { workspace = true }
+anyhow = { workspace = true }
+clap = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+toml = { workspace = true }
+walkdir = { workspace = true }
+
+[dev-dependencies]
+tempfile = { workspace = true }
